+++ CHANGES.rst
@@ -3,6 +3,8 @@
 
 - Allow exporting of plots with a categorical size attribute. [#14]
 
+- Set exported camera projection based on glue viewer state. [#13]
+
 - Add support for 2d polar scatter plots. [#12]
 
 0.3 (2021-10-19)